Fixed various issues, parser (slightly) better
Added NUMBER type to Token class, as we were regarding numbers as the
literal symbol "integer".
Fixed rule for BODY.
Renamed a few things for consistency.
Renamed first and follow sets, we probably won't need it again but
filenames with spaces in them are the Devil.
Removed extra tableDrivenParser file, already took updates from it.